+++ src/snapshot/tosvg.js
@@ -107,7 +107,7 @@ module.exports = function toSVG(gd, format) {
 
     // remove draglayer for Adobe Illustrator compatibility
     if(fullLayout._draggers) {
-        fullLayout._draggers.node().remove();
+        fullLayout._draggers.remove();
     }
 
     // in case the svg element had an explicit background color, remove this