+++ test/jasmine/tests/legend_scroll_test.js
@@ -11,7 +11,7 @@ var mock = require('../../image/mocks/legend_scroll.json');
 describe('The legend', function() {
     'use strict';
 
-    var gd, legend;
+    var gd, legend, bg;
 
     function countLegendGroups(gd) {
         return gd._fullLayout._toppaper.selectAll('g.legend').size();
@@ -27,6 +27,9 @@ describe('The legend', function() {
         return gd._fullLayout.height - gd._fullLayout.margin.t - gd._fullLayout.margin.b;
     }
 
+    function getLegendHeight() {
+        return gd._fullLayout.legend.borderwidth + getBBox(bg).height;
+    }
 
     describe('when plotted with many traces', function() {
         beforeEach(function(done) {
@@ -36,6 +39,7 @@ describe('The legend', function() {
 
             Plotly.plot(gd, mockCopy.data, mockCopy.layout).then(function() {
                 legend = document.getElementsByClassName('legend')[0];
+                bg = document.getElementsByClassName('bg')[0];
                 done();
             });
         });
@@ -43,7 +47,7 @@ describe('The legend', function() {
         afterEach(destroyGraph);
 
         it('should not exceed plot height', function() {
-            var legendHeight = getBBox(legend).height;
+            var legendHeight = getLegendHeight();
 
             expect(+legendHeight).toBe(getPlotHeight(gd));
         });
@@ -57,7 +61,7 @@ describe('The legend', function() {
 
         it('should scroll when there\'s a wheel event', function() {
             var scrollBox = legend.getElementsByClassName('scrollbox')[0],
-                legendHeight = getBBox(legend).height,
+                legendHeight = getLegendHeight(),
                 scrollBoxYMax = gd._fullLayout.legend.height - legendHeight,
                 scrollBarYMax = legendHeight -
                     constants.scrollBarHeight -
@@ -87,7 +91,7 @@ describe('The legend', function() {
 
         it('should scale the scrollbar movement from top to bottom', function() {
             var scrollBar = legend.getElementsByClassName('scrollbar')[0],
-                legendHeight = getBBox(legend).height;
+                legendHeight = getLegendHeight();
 
             // The scrollbar is 20px tall and has 4px margins
 
@@ -111,10 +115,10 @@ describe('The legend', function() {
         });
 
         it('should resize when relayout\'ed with new height', function(done) {
-            var origLegendHeight = getBBox(legend).height;
+            var origLegendHeight = getLegendHeight();
 
             Plotly.relayout(gd, 'height', gd._fullLayout.height / 2).then(function() {
-                var legendHeight = getBBox(legend).height;
+                var legendHeight = getLegendHeight();
 
                 //legend still exists and not duplicated
                 expect(countLegendGroups(gd)).toBe(1);
@@ -164,14 +168,15 @@ describe('The legend', function() {
         });
 
         it('should resize when traces added', function(done) {
-            var origLegend = document.getElementsByClassName('legend')[0];
-            var origLegendHeight = getBBox(origLegend).height;
+            legend = document.getElementsByClassName('legend')[0];
+            bg = document.getElementsByClassName('bg')[0];
+            var origLegendHeight = getLegendHeight();
 
             Plotly.addTraces(gd, { x: [1,2,3], y: [4,3,2], name: 'Test2' }).then(function() {
-                var legend = document.getElementsByClassName('legend')[0];
-                var legendHeight = getBBox(legend).height;
+                legend = document.getElementsByClassName('legend')[0];
+                bg = document.getElementsByClassName('bg')[0];
+                var legendHeight = getLegendHeight();
 
-                // clippath resized to show new trace
                 expect(+legendHeight).toBeCloseTo(+origLegendHeight + 19, 0);
 
                 done();